+++ R/S3_nobs.R
@@ -3,7 +3,9 @@
 nobs.lm_robust <- function(object, ...) object$nobs
 
 #' @export
-nobs.lh_robust <- function(object, ...) object$nobs
+# An lh_robust fit keeps its observations on the lm_robust inside it; reading
+# the top level returned NULL, in 1.0.6 as well.
+nobs.lh_robust <- function(object, ...) object$lm_robust$nobs
 
 #' @export
 nobs.iv_robust <- function(object, ...) object$nobs